Simplify application entrypoint and main window
